Add support for different size profile image
Add an extra `profile_image_size` parameter to the $args argument that
excepts `mini`, `normal`, `bigger` and `original` as size value. It
defaults to `normal`.

Add the method get `get_profile_image_url()` method that excepts the
`profile_image_size` parameter to return an URL that corresponds with
the requested size.

Resolves: #7